requirements.txt
fiftyone==0.23.5
fiftyone_brain==0.16.1
fiftyone_db==1.1.1
Flask==3.0.3
matplotlib==3.9.2
numpy==2.1.2
opencv_python==4.10.0.84
opencv_python_headless==4.9.0.80
Pillow==10.4.0
sahi==0.11.16
scipy==1.14.1
ultralytics==8.2.102
Werkzeug==3.0.4
